+++ src/graphics/Text.hpp
@@ -4,6 +4,7 @@
 #include <SFML/Graphics.hpp>
 
 #include "Shape.hpp"
+
 class Text : public Shape {
 private:
     sf::Text text;
@@ -12,7 +13,7 @@ public:
     static sf::Font font;
 
     Text(sf::Vector2f pos, sf::String TEXT,
-         sf::Color fill_color = sf::Color::Black, unsigned int font_size = 1);
+         sf::Color fill_color = sf::Color::Black, float font_size = 1);
     virtual std::size_t getPointCount() const;
     virtual sf::Vector2f getPoint(std::size_t index) const;
     void draw(sf::RenderWindow& window);
